requirements.txt
asgiref==3.6.0
astroid==2.15.5
colorama==0.4.6
dill==0.3.6
Django==4.2.1
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
pyarrow==12.0.1
pylance==0.4.21
pylint==2.17.4
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
wrapt==1.15.0
